refactor: simplify adminRoutes by grouping adminControllers into subfolders
